requirements.txt
asgiref==3.11.1
Django==6.0.3
sqlparse==0.5.5
django-ninja==1.6.2
pyjwt==2.10.1
bcrypt==4.2.0
django-cors-headers==4.6.0
python-dotenv==1.0.1
